Encapsulate validate_header_regarding_parent inside SealedHeader
