+++ crates/primitives/Cargo.toml
@@ -55,6 +55,7 @@ tempfile.workspace = true
 thiserror.workspace = true
 zstd = { version = "0.12", features = ["experimental"] }
 ahash.workspace = true
+cfg-if = "1.0.0"
 
 # `test-utils` feature
 hash-db = { version = "~0.15", optional = true }
